+++ docs/components/example/multiline-text-field-form.tsx
@@ -109,7 +109,7 @@ export default function MultilineTextFieldForm() {
         <ActionButton type="reset" variant="neutralWeak">
           초기화
         </ActionButton>
-        <ActionButton type="submit" flexGrow={1}>
+        <ActionButton type="submit" flexGrow>
           제출
         </ActionButton>
       </HStack>