+++ packages/figma/src/entities/data/__generated__/component-sets/radio.mjs
@@ -12,15 +12,19 @@ export const metadata = {
         "Large"
       ]
     },
+    "Selected": {
+      "type": "VARIANT",
+      "variantOptions": [
+        "True",
+        "False"
+      ]
+    },
     "State": {
       "type": "VARIANT",
       "variantOptions": [
         "Enabled",
         "Pressed",
-        "Disabled",
-        "Selected",
-        "Selected-Pressed",
-        "Selected-Disabled"
+        "Disabled"
       ]
     }
   }