+++ packages/figma/src/entities/data/__generated__/component-sets/select-box.d.ts
@@ -18,13 +18,18 @@ export declare const metadata: {
         "Radio"
       ]
     },
+    "Selected": {
+      "type": "VARIANT",
+      "variantOptions": [
+        "True",
+        "False"
+      ]
+    },
     "State": {
       "type": "VARIANT",
       "variantOptions": [
         "Enabled",
-        "Pressed",
-        "Selected",
-        "Selected-Pressed"
+        "Pressed"
       ]
     }
   }