+++ packages/css/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @seed-design/css
 
+## 0.1.12
+
+### Patch Changes
+
+- 62094b6: Help Bubble의 스타일 문제를 수정합니다.
+
+  - `placement=left-*` / `placement=right-*`에서 arrow가 content와 떨어져 표시되는 문제를 수정합니다.
+
+## 0.1.10
+
+### Patch Changes
+
+- ef91c21: Bottom Sheet의 스타일 문제를 수정해요.
+
+  - Close Button에 브라우저 기본 스타일이 표시되는 문제를 수정해요.
+
+## 0.1.9
+
+### Patch Changes
+
+- 5a025b7: Switch 컴포넌트를 업데이트합니다.
+
+  - size: medium → 32, small → 16으로 rename합니다.
+    - (React) `size="medium"`으로 `32`, `size="small"`로 `16`을 사용할 수 있습니다. (deprecated)
+  - size: 24를 추가합니다.
+  - 모든 size에 대해 레이블 스타일을 추가합니다. (기존: small에만 존재)
+
+- ac35731: Chip.Root `position: relative` 속성 추가
+
+  - 이제 Chip.Toggle을 사용해도 예상치 못한 스크롤이 발생하지 않습니다.
+
+- f9041e9: `CheckSelectBox`, `RadioSelectBox`의 `label`, `description` 영역을 수정합니다.
+
+  - `span` 대신 `div`를 렌더링합니다.
+  - 기본적으로 grow하도록 만들어 Badge 등 추가 요소를 넣기 쉽게 만듭니다.
+
 ## 0.1.8
 
 ### Patch Changes