+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @seed-design/mcp
 
+## 0.1.12
+
+### Patch Changes
+
+- @seed-design/figma@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [9993e0e]
+  - @seed-design/figma@0.1.11
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [aa40f66]
+  - @seed-design/figma@0.1.10
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [5a025b7]
+- Updated dependencies [f9379e0]
+  - @seed-design/figma@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes